sceneCell minor refactor
